Move redisMock to utils. fix tests
